moved the interface to control the order creation to the web app
