+++ generate_orders/package.json
@@ -12,6 +12,7 @@
     "knex": "^0.20.1",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "pg": "^7.12.1"
+    "pg": "^7.12.1",
+    "pg-listen": "^1.5.0"
   }
 }
